+++ Magent/App/AppDelegate.swift
@@ -198,6 +198,7 @@ class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCenterDele
         let threadMenuItem = NSMenuItem()
         let threadMenu = NSMenu(title: "Thread")
         threadMenu.addItem(withTitle: "New Thread", action: #selector(SplitViewController.requestNewThread), keyEquivalent: "")
+        threadMenu.addItem(withTitle: "New Thread from Branch", action: #selector(SplitViewController.requestNewThreadFromBranch), keyEquivalent: "")
         threadMenuItem.submenu = threadMenu
         mainMenu.addItem(threadMenuItem)